refractor: move searchform code to autocomplesearchbox (#44)
