+++ react/src/components/ui/Fields/index.js
@@ -1,1 +1,2 @@
 export { default as InputTextField } from './InputTextField'
+export { default as AutoCompleteInput } from './AutoCompleteInput'
